stablediffusiongallery/pyproject.toml
[tool.poetry]
name = "stablediffusiongallery"
version = "0.1.0"
description = ""
authors = ["David Aronchick <aronchick@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
black = "^22.10.0"
requests = "^2.28.1"
APScheduler = "^3.9.1"
Flask = "^2.2.2"
"ruamel.yaml" = "^0.17.21"
gunicorn = "^20.1.0"
greenlet = "^1.1.3"
gevent = "^22.10.1"
gevent-websocket = "^0.10.1"
python-dateutil = "^2.8.2"
regex = "^2022.9.13"
pytz = "^2022.5"
jsonpickle = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
